chore: extract req interface
