+++ src/backend/interfaces/index.ts
@@ -1,0 +1,5 @@
+import type { UserContext } from '../middleware/user-context';
+
+export interface ReqLike {
+  userContext?: UserContext;
+}
